executor_up/requirements.txt
grpcio==1.37.0
grpcio-tools==1.37.0
aiddl-core==0.3.1
unified-planning==0.6.0
